Add validation requests and policies

+++ app/Http/Controllers/RepositoryController.php
@@ -4,30 +4,49 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Repository;
-use App\Http\Requests\RepositoryPostRequest;
+use App\Http\Requests\RepositoryRequest;
 
 class RepositoryController extends Controller
 {
-    public function store(RepositoryPostRequest $request){
+    public function index(){
+        return view('repositories.index', [
+            'repositories' => auth()->user()->repositories
+        ]);
+    }
 
-        $validated = $request->validated();
+    public function show(Repository $repository){
 
+        $this->authorize('pass',$repository);
+        return view('repositories.show',compact('repository'));
+    }
+
+    public function create(){
+        return view('repositories.create');
+    }
+
+    public function store(RepositoryRequest $request){
         $request->user()->repositories()->create($request->all());
         return redirect()->route('repositories.index');
     }
 
-    public function update(Request $request, Repository $repository){
+    public function edit(Repository $repository){
 
-        $request->validate([
-            'url' => 'required',
-            'description' => 'required'
-        ]);
+        $this->authorize('pass',$repository);
+
+        return view('repositories.edit',compact('repository'));
+    }
+
+    public function update(RepositoryRequest $request, Repository $repository){
+
+        $this->authorize('pass',$repository);
 
         $repository->update($request->all());
         return redirect()->route('repositories.edit',$repository);
     }
 
-    public function destroy(Request $request, Repository $repository){
+    public function destroy(Repository $repository){
+
+        $this->authorize('pass',$repository);
 
         $repository->delete($repository->id);
         return redirect()->route('repositories.index');
